Made netlogentry nicer, added in corolis code from elsewhere
Split apart, getting too complicated
