+++ TestNetLogEntry/packages.config
@@ -2,4 +2,5 @@
 <packages>
   <package id="Costura.Fody" version="1.6.2" targetFramework="net46" developmentDependency="true" />
   <package id="Fody" version="2.0.0" targetFramework="net46" developmentDependency="true" />
+  <package id="Newtonsoft.Json" version="11.0.1" targetFramework="net46" />
 </packages>
